Improve code structure.
