+++ app/src/server/server_api.rs
@@ -11,7 +11,6 @@ pub mod referral;
 pub mod team;
 pub mod workspace;
 
-use std::fmt;
 use std::path::Path;
 use std::sync::Arc;
 use std::time::Duration;
@@ -40,7 +39,8 @@ use warp_core::context_flag::ContextFlag;
 use warp_core::errors::{register_error, AnyhowErrorExt, ErrorExt};
 use warp_core::telemetry::TelemetryEvent;
 use warp_managed_secrets::client::ManagedSecretsClient;
-use warp_server_client::auth::{AuthClientImpl, EXPERIMENT_ID_HEADER};
+use warp_server_client::auth::{AuthClientImpl, AuthEvent, AuthSession, EXPERIMENT_ID_HEADER};
+use warp_server_client::base_client::BaseClient as _;
 use warpui::r#async::BoxFuture;
 use warpui::{Entity, ModelContext, SingletonEntity};
 use workspace::WorkspaceClient;
@@ -373,44 +373,6 @@ cfg_if::cfg_if! {
     }
 }
 
-/// An event related to the server API itself (and not a particular API call).
-/// Most errors should be handled in callbacks to individual APIs, rather than sent over the
-/// server API channel.
-#[derive(Clone)]
-pub enum ServerApiEvent {
-    /// We made a staging API call that was blocked, which may indicate a firewall misconfiguration.
-    StagingAccessBlocked,
-    /// The user's access token was invalid, so they need to reauth before they can make
-    /// requests to warp-server.
-    NeedsReauth,
-    /// The user's account has been disabled.
-    UserAccountDisabled,
-    /// The current bearer token was refreshed.
-    AccessTokenRefreshed {
-        #[cfg_attr(target_family = "wasm", allow(dead_code))]
-        token: String,
-    },
-    /// An IAP (Identity-Aware Proxy) challenge was received, indicating that
-    /// the cached IAP credentials are stale. The UI thread should trigger a
-    /// refresh via [`super::iap::IapManager::start_refresh`].
-    IapChallengeReceived,
-}
-
-impl fmt::Debug for ServerApiEvent {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::StagingAccessBlocked => f.write_str("StagingAccessBlocked"),
-            Self::NeedsReauth => f.write_str("NeedsReauth"),
-            Self::UserAccountDisabled => f.write_str("UserAccountDisabled"),
-            Self::AccessTokenRefreshed { .. } => f
-                .debug_struct("AccessTokenRefreshed")
-                .field("token", &"<redacted>")
-                .finish(),
-            Self::IapChallengeReceived => f.write_str("IapChallengeReceived"),
-        }
-    }
-}
-
 /// An API wrapper struct with methods to requests to warp-server.
 ///
 /// Prefer NOT adding new methods directly on this struct; instead, add to one of the existing
@@ -419,12 +381,11 @@ impl fmt::Debug for ServerApiEvent {
 pub struct ServerApi {
     client: Arc<http_client::Client>,
     auth_state: Arc<AuthState>,
-    event_sender: async_channel::Sender<ServerApiEvent>,
+    event_sender: async_channel::Sender<AuthEvent>,
+    auth_session: Arc<AuthSession>,
     // TODO(jeff): Make `TelemetryApi` another type of client, and move it off `ServerApi`.
     telemetry_api: TelemetryApi,
     last_server_time: Arc<Mutex<Option<ServerTime>>>,
-    // We technically use OAuth2 for headless device authentication.
-    oauth_client: self::auth::OAuth2Client,
     /// Cached ambient workload token for requests from ambient agents.
     ambient_workload_token: Arc<Mutex<Option<warp_isolation_platform::WorkloadToken>>>,
     /// The ambient agent task ID for requests from cloud agents.
@@ -441,13 +402,18 @@ pub struct ServerApi {
 impl ServerApi {
     fn new(
         auth_state: Arc<AuthState>,
-        event_sender: async_channel::Sender<ServerApiEvent>,
+        event_sender: async_channel::Sender<AuthEvent>,
         agent_source: Option<ai::AgentSource>,
         iap_state: Option<Arc<IapState>>,
+        ctx: &mut ModelContext<ServerApiProvider>,
     ) -> Self {
         let mut client = http_client::Client::new();
         if let Some(state) = iap_state.as_ref() {
             client.set_iap_token_provider(state.clone());
+        }
+        let mut telemetry_api = TelemetryApi::new();
+        if ContextFlag::NetworkLogConsole.is_enabled() {
+            super::network_logging::init([&mut client, &mut telemetry_api.client], ctx);
         }
         Self::new_with_parts(
             Arc::new(client),
@@ -455,15 +421,17 @@ impl ServerApi {
             event_sender,
             agent_source,
             iap_state,
+            telemetry_api,
         )
     }
 
     fn new_with_parts(
         client: Arc<http_client::Client>,
         auth_state: Arc<AuthState>,
-        event_sender: async_channel::Sender<ServerApiEvent>,
+        event_sender: async_channel::Sender<AuthEvent>,
         agent_source: Option<ai::AgentSource>,
         iap_state: Option<Arc<IapState>>,
+        telemetry_api: TelemetryApi,
     ) -> Self {
         // We generate a random user ID for evals so we can run evals in parallel.
         #[cfg(feature = "agent_mode_evals")]
@@ -472,15 +440,19 @@ impl ServerApi {
             Some(EVAL_USER_IDS[rand::thread_rng().gen_range(0..EVAL_USER_IDS.len())])
         };
 
-        let oauth_client = Self::create_oauth_client();
+        let auth_session = Arc::new(AuthSession::new(
+            client.clone(),
+            auth_state.clone(),
+            event_sender.clone(),
+        ));
 
         Self {
             client,
             auth_state,
             event_sender,
-            telemetry_api: TelemetryApi::new(),
+            auth_session,
+            telemetry_api,
             last_server_time: Arc::new(Mutex::new(None)),
-            oauth_client,
             ambient_workload_token: Arc::new(Mutex::new(None)),
             ambient_agent_task_id: Arc::new(RwLock::new(None)),
             agent_source,
@@ -496,13 +468,13 @@ impl ServerApi {
         let auth_state = Arc::new(AuthState::new_for_test());
         let client = Arc::new(http_client::Client::new_for_test());
 
-        Self::new_with_parts(client, auth_state, tx, None, None)
+        Self::new_with_parts(client, auth_state, tx, None, None, TelemetryApi::new())
     }
 
-    #[cfg(test)]
+    #[cfg(all(test, feature = "skip_login"))]
     fn new_for_test_with_bearer_token(
         bearer_token: Option<String>,
-        event_sender: async_channel::Sender<ServerApiEvent>,
+        event_sender: async_channel::Sender<AuthEvent>,
     ) -> Self {
         let auth_state = Arc::new(AuthState::new_logged_out_for_test());
         if let Some(bearer_token) = bearer_token {
@@ -514,13 +486,16 @@ impl ServerApi {
             event_sender,
             None,
             None,
+            TelemetryApi::new(),
         )
     }
 
     pub fn allowed_to_refresh_token(&self) -> bool {
-        self.auth_state
-            .credentials()
-            .is_none_or(|credentials| !credentials.is_externally_managed())
+        self.auth_session.allowed_to_refresh_token()
+    }
+
+    pub async fn get_or_refresh_access_token(&self) -> Result<crate::auth::credentials::AuthToken> {
+        self.auth_session.get_or_refresh_access_token().await
     }
 
     fn access_token_ignoring_validity(&self) -> Option<String> {
@@ -556,10 +531,7 @@ impl ServerApi {
                 "Received IAP challenge (status {}); notifying IapManager",
                 response.status()
             );
-            if let Err(err) = self
-                .event_sender
-                .try_send(ServerApiEvent::IapChallengeReceived)
-            {
+            if let Err(err) = self.event_sender.try_send(AuthEvent::IapChallengeReceived) {
                 log::warn!("Failed to enqueue IapChallengeReceived event: {err}");
             }
             true
@@ -590,7 +562,7 @@ impl ServerApi {
                     log::warn!(
                         "Received IAP challenge on eventsource (status {status}); notifying IapManager"
                     );
-                    if let Err(err) = event_sender.try_send(ServerApiEvent::IapChallengeReceived) {
+                    if let Err(err) = event_sender.try_send(AuthEvent::IapChallengeReceived) {
                         log::warn!(
                             "Failed to enqueue IapChallengeReceived event from eventsource: {err}"
                         );
@@ -639,23 +611,6 @@ impl ServerApi {
         headers.retain(|(name, _)| *name != CLOUD_AGENT_ID_HEADER);
         headers.push((CLOUD_AGENT_ID_HEADER, task_id.to_string()));
         Ok(headers)
-    }
-
-    fn create_oauth_client() -> self::auth::OAuth2Client {
-        let server_root =
-            Url::parse(&ChannelState::server_root_url()).expect("Server root URL must be valid");
-
-        let token_url = server_root
-            .join("/api/v1/oauth/token")
-            .expect("Invalid token URL");
-
-        let device_url = server_root
-            .join("/api/v1/oauth/device/auth")
-            .expect("Invalid device URL");
-
-        oauth2::basic::BasicClient::new(oauth2::ClientId::new("warp-cli".to_string()))
-            .set_token_uri(oauth2::TokenUrl::from_url(token_url))
-            .set_device_authorization_url(oauth2::DeviceAuthorizationUrl::from_url(device_url))
     }
 
     pub fn send_graphql_request<'a, QF, O: warp_graphql::client::Operation<QF> + Send + 'a>(
@@ -1539,25 +1494,19 @@ impl ServerApiProvider {
     ) -> Self {
         let (event_sender, event_receiver) = async_channel::bounded(10);
 
-        let mut server_api =
-            ServerApi::new(auth_state.clone(), event_sender, agent_source, iap_state);
-
-        if ContextFlag::NetworkLogConsole.is_enabled() {
-            super::network_logging::init(
-                [
-                    Arc::get_mut(&mut server_api.client)
-                        .expect("guaranteed there is only one copy of client"),
-                    &mut server_api.telemetry_api.client,
-                ],
-                ctx,
-            );
-        }
+        let server_api = ServerApi::new(
+            auth_state.clone(),
+            event_sender,
+            agent_source,
+            iap_state,
+            ctx,
+        );
 
         ctx.spawn_stream_local(
             event_receiver,
             move |_, event, ctx| {
                 match event {
-                    ServerApiEvent::UserAccountDisabled => {
+                    AuthEvent::UserAccountDisabled => {
                         // We dispatch a global action here because the log out code requires
                         // `server_api`, causing a circular model reference panic when it calls
                         // `ServerApiProvider` to get access.
@@ -1565,7 +1514,7 @@ impl ServerApiProvider {
                         // to events; it's prone to these sorts of circular reference issues.
                         ctx.dispatch_global_action("app:log_out", ());
                     }
-                    ServerApiEvent::NeedsReauth => {
+                    AuthEvent::NeedsReauth => {
                         // AuthManager depends on a reference to ServerApi, so ServerApi can't easily
                         // hold a ref to AuthManager. To get around this, we emit an event on ServerApi
                         // and handle calling the AuthManager here instead.
@@ -1573,7 +1522,7 @@ impl ServerApiProvider {
                             auth_manager.set_needs_reauth(true, ctx);
                         });
                     }
-                    ServerApiEvent::IapChallengeReceived => {
+                    AuthEvent::IapChallengeReceived => {
                         IapManager::handle(ctx)
                             .update(ctx, |manager, ctx| manager.handle_challenge(ctx));
                     }
@@ -1586,7 +1535,10 @@ impl ServerApiProvider {
             |_, _| {},
         );
         let server_api = Arc::new(server_api);
-        let auth_client = Arc::new(AuthClientImpl::new(server_api.clone()));
+        let auth_client = Arc::new(AuthClientImpl::new(
+            server_api.clone(),
+            server_api.auth_session.clone(),
+        ));
         Self {
             server_api,
             auth_client,
@@ -1610,7 +1562,10 @@ impl ServerApiProvider {
     #[cfg(test)]
     pub fn new_for_test() -> Self {
         let server_api = Arc::new(ServerApi::new_for_test());
-        let auth_client = Arc::new(AuthClientImpl::new(server_api.clone()));
+        let auth_client = Arc::new(AuthClientImpl::new(
+            server_api.clone(),
+            server_api.auth_session.clone(),
+        ));
         Self {
             server_api,
             auth_client,
@@ -1672,224 +1627,7 @@ impl ServerApiProvider {
 }
 
 impl Entity for ServerApiProvider {
-    type Event = ServerApiEvent;
+    type Event = AuthEvent;
 }
 
 impl SingletonEntity for ServerApiProvider {}
-
-#[cfg(test)]
-mod tests {
-    use std::borrow::Cow;
-    use std::future::Future;
-    use std::pin::Pin;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-
-    use cynic::{GraphQlError, GraphQlResponse};
-    use futures::executor::block_on;
-
-    use super::*;
-    use crate::server::graphql::GraphQLError;
-
-    struct FakeGraphqlOperation {
-        expected_auth_token: Option<String>,
-        send_count: Arc<AtomicUsize>,
-        result: FakeGraphqlResult,
-    }
-
-    enum FakeGraphqlResult {
-        Success,
-        Rejected(StatusCode),
-        ResponseErrors(Vec<String>),
-    }
-
-    impl FakeGraphqlOperation {
-        fn successful(expected_auth_token: Option<&str>, send_count: Arc<AtomicUsize>) -> Self {
-            Self {
-                expected_auth_token: expected_auth_token.map(ToOwned::to_owned),
-                send_count,
-                result: FakeGraphqlResult::Success,
-            }
-        }
-
-        fn rejected(
-            expected_auth_token: Option<&str>,
-            send_count: Arc<AtomicUsize>,
-            status: StatusCode,
-        ) -> Self {
-            Self {
-                expected_auth_token: expected_auth_token.map(ToOwned::to_owned),
-                send_count,
-                result: FakeGraphqlResult::Rejected(status),
-            }
-        }
-
-        fn response_errors(
-            expected_auth_token: Option<&str>,
-            send_count: Arc<AtomicUsize>,
-            messages: Vec<String>,
-        ) -> Self {
-            Self {
-                expected_auth_token: expected_auth_token.map(ToOwned::to_owned),
-                send_count,
-                result: FakeGraphqlResult::ResponseErrors(messages),
-            }
-        }
-    }
-
-    impl warp_graphql::client::Operation<()> for FakeGraphqlOperation {
-        fn operation_name(&self) -> Option<Cow<'_, str>> {
-            Some(Cow::Borrowed("FakeGraphqlOperation"))
-        }
-
-        fn send_request(
-            self,
-            _client: Arc<http_client::Client>,
-            options: warp_graphql::client::RequestOptions,
-        ) -> Pin<
-            Box<
-                dyn Future<Output = std::result::Result<GraphQlResponse<()>, GraphQLError>>
-                    + Send
-                    + 'static,
-            >,
-        >
-        where
-            Self: Sized,
-        {
-            Box::pin(async move {
-                assert_eq!(options.auth_token, self.expected_auth_token);
-                self.send_count.fetch_add(1, Ordering::SeqCst);
-                match self.result {
-                    FakeGraphqlResult::Success => Ok(GraphQlResponse {
-                        data: Some(()),
-                        errors: None,
-                    }),
-                    FakeGraphqlResult::Rejected(status) => Err(GraphQLError::HttpError {
-                        status,
-                        body: "redacted auth rejection".to_string(),
-                    }),
-                    FakeGraphqlResult::ResponseErrors(messages) => Ok(GraphQlResponse {
-                        data: None,
-                        errors: Some(
-                            messages
-                                .into_iter()
-                                .map(|message| GraphQlError::new(message, None, None, None))
-                                .collect(),
-                        ),
-                    }),
-                }
-            })
-        }
-    }
-
-    fn has_error_message(error: &anyhow::Error, expected: &str) -> bool {
-        error.chain().any(|cause| cause.to_string() == expected)
-    }
-
-    #[test]
-    fn send_graphql_request_refresh_enabled_uses_auth_state() {
-        let server_api = ServerApi::new_for_test();
-        let send_count = Arc::new(AtomicUsize::new(0));
-
-        block_on(server_api.send_graphql_request(
-            FakeGraphqlOperation::successful(None, send_count.clone()),
-            None,
-        ))
-        .unwrap();
-
-        assert!(server_api.allowed_to_refresh_token());
-        assert_eq!(send_count.load(Ordering::SeqCst), 1);
-    }
-
-    #[test]
-    fn send_graphql_request_refresh_disabled_uses_provided_bearer_token() {
-        let (event_sender, _) = async_channel::unbounded();
-        let server_api = ServerApi::new_for_test_with_bearer_token(
-            Some("daemon-token".to_string()),
-            event_sender,
-        );
-        let send_count = Arc::new(AtomicUsize::new(0));
-
-        block_on(server_api.send_graphql_request(
-            FakeGraphqlOperation::successful(Some("daemon-token"), send_count.clone()),
-            None,
-        ))
-        .unwrap();
-
-        assert!(!server_api.allowed_to_refresh_token());
-        assert_eq!(send_count.load(Ordering::SeqCst), 1);
-    }
-
-    #[test]
-    fn send_graphql_request_refresh_disabled_missing_token_returns_auth_error() {
-        let (event_sender, event_receiver) = async_channel::unbounded();
-        let server_api = ServerApi::new_for_test_with_bearer_token(None, event_sender);
-        let send_count = Arc::new(AtomicUsize::new(0));
-
-        let error = block_on(server_api.send_graphql_request(
-            FakeGraphqlOperation::successful(Some("unused-token"), send_count.clone()),
-            None,
-        ))
-        .unwrap_err();
-
-        assert!(has_error_message(
-            &error,
-            "missing authentication credentials"
-        ));
-        assert_eq!(send_count.load(Ordering::SeqCst), 0);
-        assert!(event_receiver.try_recv().is_err());
-    }
-
-    #[test]
-    fn send_graphql_request_refresh_disabled_auth_rejection_is_credentials_rejected() {
-        let (event_sender, event_receiver) = async_channel::unbounded();
-        let server_api = ServerApi::new_for_test_with_bearer_token(
-            Some("daemon-token".to_string()),
-            event_sender,
-        );
-        let send_count = Arc::new(AtomicUsize::new(0));
-
-        let error = block_on(server_api.send_graphql_request(
-            FakeGraphqlOperation::rejected(
-                Some("daemon-token"),
-                send_count.clone(),
-                StatusCode::UNAUTHORIZED,
-            ),
-            None,
-        ))
-        .unwrap_err();
-
-        assert!(has_error_message(
-            &error,
-            "server rejected authentication credentials"
-        ));
-        assert_eq!(send_count.load(Ordering::SeqCst), 1);
-        assert!(event_receiver.try_recv().is_err());
-    }
-
-    #[test]
-    fn send_graphql_request_refresh_disabled_user_not_in_context_is_credentials_rejected() {
-        let (event_sender, event_receiver) = async_channel::unbounded();
-        let server_api = ServerApi::new_for_test_with_bearer_token(
-            Some("daemon-token".to_string()),
-            event_sender,
-        );
-        let send_count = Arc::new(AtomicUsize::new(0));
-
-        let error = block_on(server_api.send_graphql_request(
-            FakeGraphqlOperation::response_errors(
-                Some("daemon-token"),
-                send_count.clone(),
-                vec!["User not in context: Not found".to_string()],
-            ),
-            None,
-        ))
-        .unwrap_err();
-
-        assert!(has_error_message(
-            &error,
-            "server rejected authentication credentials"
-        ));
-        assert_eq!(send_count.load(Ordering::SeqCst), 1);
-        assert!(event_receiver.try_recv().is_err());
-    }
-}
