+++ app/src/server/server_api/auth.rs
@@ -1,243 +1,12 @@
-use std::result::Result as StdResult;
-use std::sync::Arc;
-
-use anyhow::{bail, Result};
-use firebase::FetchAccessTokenResponse;
-use instant::Duration;
 use thiserror::Error;
-use warp_graphql::queries::get_user::UserOutput as GqlUserOutput;
 #[cfg(test)]
 pub use warp_server_client::auth::MockAuthClient;
 pub use warp_server_client::auth::{
     AuthClient, FetchUserResult, MintCustomTokenError, SyncedUserSettings, UserAuthenticationError,
 };
-use warpui::r#async::BoxFuture;
-
-use super::ServerApi;
-use crate::auth::credentials::{AuthToken, Credentials, FirebaseToken, LoginToken, RefreshToken};
-use crate::auth::user::{FirebaseAuthTokens, User};
-use crate::auth::UserUid;
-use crate::channel::ChannelState;
-use crate::convert_to_server_experiment;
-use crate::server::experiments::ServerExperiment;
-use crate::server::server_api::ServerApiEvent;
-
-const FETCH_ACCESS_TOKEN_TIMEOUT: Duration = Duration::from_secs(5);
-
-impl ServerApi {
-    pub(super) async fn access_token(&self) -> Result<AuthToken> {
-        if cfg!(feature = "skip_login") {
-            bail!("skip_login enabled; failing all authenticated requests");
-        }
-
-        let Some(credentials) = self.auth_state.credentials() else {
-            bail!("missing authentication credentials");
-        };
-
-        match credentials {
-            Credentials::ApiKey { key, .. } => Ok(AuthToken::ApiKey(key)),
-            Credentials::Bearer(token) => Ok(AuthToken::Bearer(token)),
-            Credentials::Firebase(auth_tokens) => {
-                let expiration_time = auth_tokens.expiration_time;
-
-                // Generate a new ID token if the token has expired or will expire in the
-                // next five minutes. This matches the behavior of the Firebase Auth SDK.
-                if chrono::Local::now().fixed_offset() + chrono::Duration::minutes(5)
-                    >= expiration_time
-                {
-                    let refresh_token = auth_tokens.refresh_token.clone();
-                    let firebase_token = FirebaseToken::Refresh(RefreshToken::new(refresh_token));
-
-                    let result = fetch_auth_tokens(self.client.clone(), firebase_token).await;
-
-                    if let Err(UserAuthenticationError::DeniedAccessToken(_)) = result {
-                        let _ = self.event_sender.send(ServerApiEvent::NeedsReauth).await;
-                    }
-                    let new_firebase_token_info = result?;
-                    self.auth_state
-                        .update_firebase_tokens(new_firebase_token_info.clone());
-                    let _ = self
-                        .event_sender
-                        .send(ServerApiEvent::AccessTokenRefreshed {
-                            token: new_firebase_token_info.id_token.clone(),
-                        })
-                        .await;
-                    return Ok(AuthToken::Firebase(new_firebase_token_info.id_token));
-                }
-
-                Ok(AuthToken::Firebase(auth_tokens.id_token))
-            }
-            Credentials::SessionCookie => Ok(AuthToken::NoAuth),
-            #[cfg(any(feature = "integration_tests", feature = "skip_login"))]
-            Credentials::Test => Ok(AuthToken::NoAuth),
-        }
-    }
-
-    pub async fn get_or_refresh_access_token(&self) -> Result<AuthToken> {
-        self.access_token().await
-    }
-}
-
-/// Exchange a long-lived token for fresh [`Credentials`].
-pub(super) async fn exchange_credentials(
-    client: Arc<http_client::Client>,
-    token: LoginToken,
-) -> StdResult<Credentials, UserAuthenticationError> {
-    match token {
-        LoginToken::Firebase(firebase_token) => {
-            let tokens = fetch_auth_tokens(client, firebase_token).await?;
-            Ok(Credentials::Firebase(tokens))
-        }
-        LoginToken::ApiKey(key) => Ok(Credentials::ApiKey {
-            key,
-            owner_type: None,
-        }),
-        LoginToken::SessionCookie => Ok(Credentials::SessionCookie),
-    }
-}
-
-fn fetch_auth_tokens(
-    client: Arc<http_client::Client>,
-    token: FirebaseToken,
-) -> BoxFuture<'static, StdResult<FirebaseAuthTokens, UserAuthenticationError>> {
-    Box::pin(async move {
-        let firebase_api_key = ChannelState::firebase_api_key();
-        let url = token.access_token_url(&firebase_api_key);
-        let request_body = token.access_token_request_body();
-        let proxy_url = token.proxy_url(&ChannelState::server_root_url(), &firebase_api_key);
-        let response = match client
-            .post(&url)
-            .form(&request_body)
-            .timeout(FETCH_ACCESS_TOKEN_TIMEOUT)
-            .send()
-            .await
-        {
-            Ok(response) => match response.error_for_status_ref() {
-                Ok(_) => Ok(response),
-                Err(error) => {
-                    log::warn!(
-                        "Request to firebase to fetch access token completed, but was unsuccessful: {error:?}"
-                    );
-
-                    fetch_access_token_via_proxy(client, &request_body, proxy_url).await
-                }
-            },
-            Err(error) => {
-                log::warn!("Failed to make response to firebase to fetch access token: {error:?}");
-
-                fetch_access_token_via_proxy(client, &request_body, proxy_url).await
-            }
-        }?;
-
-        let response = response
-            .json::<FetchAccessTokenResponse>()
-            .await
-            .map_err(anyhow::Error::from)?;
-        match response {
-            FetchAccessTokenResponse::Success {
-                id_token,
-                expires_in,
-                refresh_token,
-            } => Ok(FirebaseAuthTokens::from_response(
-                id_token,
-                refresh_token,
-                expires_in,
-            )?),
-            FetchAccessTokenResponse::Error { error } => Err(error.into()),
-        }
-    })
-}
-
-fn fetch_access_token_via_proxy<'a>(
-    client: Arc<http_client::Client>,
-    request_body: &'a [(&'a str, &'a str)],
-    proxy_url: String,
-) -> BoxFuture<'a, Result<http_client::Response>> {
-    Box::pin(async move {
-        client
-            .post(&proxy_url)
-            .form(request_body)
-            .send()
-            .await
-            .map_err(anyhow::Error::from)
-    })
-}
-
-/// The [`oauth2::Client`] type, specialized to the endpoints that we require.
-pub type OAuth2Client = oauth2::basic::BasicClient<
-    oauth2::EndpointNotSet, // HasAuthUrl
-    oauth2::EndpointSet,    // HasDeviceAuthUrl
-    oauth2::EndpointNotSet, // HasIntrospectionUrl
-    oauth2::EndpointNotSet, // HasRevocationUrl
-    oauth2::EndpointSet,    // HasTokenUrl
->;
-
-/// Intermediate type produced by converting a [`GqlUserOutput`] from the server.
-pub(crate) struct UserProperties {
-    pub(crate) user: User,
-    pub(crate) server_experiments: Vec<ServerExperiment>,
-    pub(crate) llms: crate::ai::llms::ModelsByFeature,
-}
-
-impl From<GqlUserOutput> for UserProperties {
-    fn from(user_output: GqlUserOutput) -> Self {
-        let principal_type = user_output
-            .principal_type
-            .map(|pt| pt.into())
-            .unwrap_or_default();
-        let user_properties = user_output.user;
-
-        let is_on_work_domain = user_properties.is_on_work_domain;
-        let is_onboarded = user_properties.is_onboarded;
-        let global_skills = user_properties.global_skills;
-
-        let linked_at = user_properties
-            .anonymous_user_info
-            .as_ref()
-            .and_then(|info| info.linked_at);
-
-        let anonymous_user_type = user_properties
-            .anonymous_user_info
-            .as_ref()
-            .map(|info| info.anonymous_user_type.clone());
-        let personal_object_limits = user_properties
-            .anonymous_user_info
-            .and_then(|info| info.personal_object_limits.clone());
-        let user_profile = user_properties.profile;
-        let local_id = UserUid::new(user_profile.uid.as_str());
-        let needs_sso_link = user_profile.needs_sso_link;
-
-        let server_experiments: Vec<ServerExperiment> = user_properties
-            .experiments
-            .and_then(|experiments| convert_to_server_experiment!(experiments))
-            .unwrap_or_default();
-
-        // Convert LLM model choices from GraphQL response
-        let llms = user_properties.llms.try_into().unwrap_or_default();
-
-        let user = User {
-            is_onboarded,
-            local_id,
-            metadata: user_profile.into(),
-            needs_sso_link,
-            anonymous_user_type: anonymous_user_type.and_then(|t| t.try_into().ok()),
-            is_on_work_domain,
-            linked_at,
-            personal_object_limits: personal_object_limits.and_then(|t| t.try_into().ok()),
-            principal_type,
-            global_skills,
-        };
-
-        UserProperties {
-            user,
-            server_experiments,
-            llms,
-        }
-    }
-}
 
 #[derive(Error, Debug)]
-/// Error type when creating anonymous users
+/// Error type when creating anonymous users.
 pub enum AnonymousUserCreationError {
     #[error("The network request to create the anonymous user failed")]
     CreationFailed,
