+++ app/src/server/server_api/auth_tests.rs
@@ -45,7 +45,7 @@ fn access_token_skip_login_rejects_bearer_token() {
     let server_api =
         ServerApi::new_for_test_with_bearer_token(Some("daemon-token".to_string()), event_sender);
 
-    let error = futures::executor::block_on(server_api.access_token()).unwrap_err();
+    let error = futures::executor::block_on(server_api.get_or_refresh_access_token()).unwrap_err();
 
     assert_eq!(
         error.to_string(),
