+++ crates/warp_server_client/src/graphql_helpers.rs
@@ -98,3 +98,7 @@ where
         })
     })
 }
+
+#[cfg(test)]
+#[path = "graphql_helpers_tests.rs"]
+mod tests;
